chore: move inline-script to src
