+++ bin/build-inline-script.js
@@ -13,7 +13,7 @@ const writeFile = promisify(fs.writeFile)
 const themeColors = fromPairs(themes.map(_ => ([_.name, _.color])))
 
 export async function buildInlineScript () {
-  let inlineScriptPath = path.join(__dirname, '../inline-script.js')
+  let inlineScriptPath = path.join(__dirname, '../src/inline-script/inline-script.js')
 
   let bundle = await rollup({
     input: inlineScriptPath,
@@ -38,8 +38,8 @@ export async function buildInlineScript () {
   let fullCode = `${code}//# sourceMappingURL=/inline-script.js.map`
   let checksum = crypto.createHash('sha256').update(fullCode).digest('base64')
 
-  await writeFile(path.resolve(__dirname, '../inline-script-checksum.json'),
-    JSON.stringify({ checksum }), 'utf8')
+  await writeFile(path.resolve(__dirname, '../src/inline-script/checksum.js'),
+    `module.exports = ${JSON.stringify(checksum)}`, 'utf8')
   await writeFile(path.resolve(__dirname, '../static/inline-script.js.map'),
     map.toString(), 'utf8')
